--- populate/src/main/resources/seed/portals/demo/studies/heartdemo/enrollees/jtsalk.json
@@ -1,20 +1,0 @@
-{
-  "linkedUsername": "jtsalk@test.com",
-  "shortcode": "HDJSLK",
-  "simulateSubmissions": true,
-  "submittedHoursAgo": 52,
-  "preEnrollmentResponseDto": {
-    "qualified": true,
-    "surveyStableId": "hd_hd_preenroll",
-    "surveyVersion": 1,
-    "answers": [
-      {"questionStableId":"hd_hd_preenroll_southAsianAncestry", "stringValue": "yes", "viewedLanguage": "dev"},
-      {"questionStableId":"hd_hd_preenroll_understandsEnglish", "stringValue": "yes", "viewedLanguage": "en"},
-      {"questionStableId":"hd_hd_preenroll_isAdult", "stringValue": "yes", "viewedLanguage": "es"},
-      {"questionStableId":"hd_hd_preenroll_livesInUS", "stringValue": "yes", "viewedLanguage": "es"},
-      {"questionStableId":"qualified", "booleanValue": true}
-    ]
-  },
-  "surveyResponseDtos": [],
-  "participantNoteDtos": []
-}
